correction of a bug in check_dmqc.m that appears in recent Matlab version

+++ src/create_dm_files/check_dmqc.m
@@ -162,7 +162,7 @@ if isfield(CHECK,param)
                 case{'OW'}
                     set(p,'FaceColor',[1 0.8 0.8],'FaceAlpha',0.5,'EdgeColor',[0.9 0.9 0.9]);
             end
-            text(xpos,ypos,thecorrection,'interpreter','none')
+            text(double(xpos),double(ypos),thecorrection,'interpreter','none')
         end
         clear xlim ylim
